requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
pandas==2.1.3
numpy==1.26.2
pyyaml==6.0.1
requests==2.31.0
python-dotenv==1.0.0
lxml==4.9.3
html5lib==1.1
Flask==2.3.3
gunicorn==21.2.0
APScheduler==3.10.4
Flask-SQLAlchemy==3.1.1
